+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Add `gar_scope_config` 
 * Add Docker build available at `gcr.io/gcer-public/googleauthr`
+* More API error feedback if using googleAuthR.verbose < 3
+* Auto-auth should succeed if one of the scopes needed is present
 
 # googleAuthR v1.1.1